Adds: rename subdir from source to DreamAN
