+++ macros/mainDVCS2.C
@@ -5,11 +5,11 @@
 #include "TCanvas.h"
 #include <optional>
 
-#include "../source/ParticleInformation/RECParticle.h"
-#include "../source/Cuts/ElectronCut.h"
-#include "../source/DrawHist/DrawAndSave.h"
-#include "../source/core/FilesInPath.h"
-#include "../source/Cuts/EventCut.h"
+#include "../DreamAN/ParticleInformation/RECParticle.h"
+#include "../DreamAN/Cuts/ElectronCut.h"
+#include "../DreamAN/DrawHist/DrawAndSave.h"
+#include "../DreamAN/core/FilesInPath.h"
+#include "../DreamAN/Cuts/EventCut.h"
 
 using namespace ROOT;
 using namespace ROOT::RDF;
