+++ macros/testRDF.C
@@ -4,8 +4,8 @@
 #include "RHipoDS.hxx"
 #include "TCanvas.h"
 
-#include "../source/ParticleInformation/RECParticle.h"
-#include "../source/Cuts/ElectronCut.h"
+#include "../DreamAN/ParticleInformation/RECParticle.h"
+#include "../DreamAN/Cuts/ElectronCut.h"
 
 using namespace ROOT;
 using namespace ROOT::RDF;
